+++ lib/traitor/config.rb
@@ -4,22 +4,21 @@ module Traitor
     @create_kwargs  = {}
     @build_kwargs = {}
     @build_with_list = false # use (**attributes.merge(build_kwargs)) instead of (attributes, **build_kwargs)
+    @no_callbacks = false
 
     class << self
-      attr_accessor :create_method, :create_kwargs, :build_kwargs, :build_with_list
+      attr_accessor :create_method, :create_kwargs, :build_kwargs, :build_with_list, :no_callbacks
 
       def configure_for_rails!
         @create_method  = :save
         @create_kwargs  = { validate: false }
         @build_kwargs = { without_protection: true }
-        @build_with_list = false
       end
 
       def configure_safe_for_rails!
         @create_method  = :save
         @create_kwargs  = {}
         @build_kwargs = {}
-        @build_with_list = false
       end
 
       def reset!
@@ -27,6 +26,19 @@ module Traitor
         @create_kwargs = {}
         @build_kwargs = {}
         @build_with_list = false
+        @no_callbacks = false
+      end
+
+      def stash!
+        @old_config = Hash[
+          self.instance_variables.map { |att| [att, self.instance_variable_get(att)] }
+        ]
+      end
+
+      def restore!
+        return unless @old_config
+        @old_config.each { |att, val| self.instance_variable_set(att, val) }
+        @old_config = nil
       end
     end
   end