+++ spec/spec_helper.rb
@@ -9,21 +9,28 @@ require 'pry'
 require 'timecop'
 
 require 'traitor'
+require 'traitor/helpers/rspec' # allow traitor_# config metadata
 
 RSpec.configure do |config|
   config.expect_with :rspec do |expectations|
     expectations.syntax = :expect
+  end
+
+  config.after do
+    Traitor.reset!
   end
 end
 
 # for testing
 
 class TestClass
-  attr_accessor :param1, :param2
+  attr_accessor :param1, :param2, :args, :kwargs
   def initialize(*args, **kwargs)
     attrs = args[0] || kwargs
     @param1 = attrs[:param1]
     @param2 = attrs[:param2]
+    @args = args
+    @kwargs = kwargs
   end
 
   def create(*args, **kwargs)